requirements.txt
instaloader
shazamio
pandas
python-dotenv
selenium
spotipy
python-dotenv